brought image picker directly into add new workout
