+++ src/components/ExerciseListComponent/ExerciseListComponent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Text, View, Image } from 'react-native';
 import { MinutesAndSeconds } from '../../HelperFunctions';
 
 const ExerciseListComponent = (props) => {
@@ -15,6 +15,9 @@ const ExerciseListComponent = (props) => {
     <View key={index} style={styles.exerciseBox}>
       <View style={styles.exerciseHeader}>
         <Text style={styles.excerciseText}> {exercise.title} </Text>
+        <View style={styles.exerciseImage}>
+        {exercise.image && <Image source={{ uri: exercise.image }} style={{ width: 75, height: 75 }} />}
+        </View>
         <View style={{margin: 8}}>
           <MinutesAndSeconds seconds={exercise.duration} />
         </View>
@@ -25,6 +28,10 @@ const ExerciseListComponent = (props) => {
 };
 
 const styles = StyleSheet.create({
+  exerciseImage: {
+    justifyContent: 'center',
+    margin: 5,
+  },
   exerciseBox: {
     margin: 20,
     backgroundColor: 'white'
