+++ src/components/ImagePickerComponent/ImagePickerComponent.js
@@ -3,28 +3,8 @@ import { Button, Image, View, Platform } from 'react-native';
 import * as ImagePicker from 'expo-image-picker';
 
 const ImagePickerComponent = () => {
-  const [image, setImage] = useState(null);
 
-  const pickImage = async () => {
-    let result = await ImagePicker.launchImageLibraryAsync({
-      mediaTypes: ImagePicker.MediaTypeOptions.Images,
-      allowsEditing: true,
-      quality: 1,
-    });
-
-    console.log(result);
-
-    if (!result.cancelled) {
-      setImage(result.uri);
-    }
-  };
-
-  return (
-  <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-    <Button title="Pick an image from camera roll" onPress={pickImage} />
-    {image && <Image source={{ uri: image }} style={{ width: 200, height: 200 }} />}
-  </View>
-  );
+  
 }
 
 export { ImagePickerComponent }
